pyproject.toml
[project]
name = "mpris-download-music"
version = "0.1.0"
description = "Download current playing music using mpris and spotdl"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "googlesearch-python>=1.3.0",
    "spotdl>=4.2.11",
]
